Update roadmap.md
Line 8: If you give a description for all the other design values, you should there too.
Line 10: I don't know what this design value means, and your description of it is also a bit confusing. I understand what you're trying to say about your app (that it should be as seamless as its name) but the wording is a bit off.
Line 12: It might sound better to say something like "When you're using Seamless, you get to focus on solving your problems while we focus on solving ours." Although, it's probably not great to say Seamless has problems. I know what you're trying to say but there's probably a better way to phrase it. What that is, I don't know because I'm not a tech person, but it's something to think about.
Line 33: I don't really know what you're talking about, but it seems like it may be out of place or unimportant.
Line 57: I added "what's missing in today's world" as a place holder. You should specify like "development world" or "design landscape" or whatever, I just don't have the tech language to determine what the right word would be
Line 62: I don't know exactly what this is saying but "as soon as we are confident" is not the strongest word choice, you should be confident in your own product.