Make sure stderr and stdout are being consumed and don't block the diffing on Windows
The problem lied in that stdout and stderr streams not being consumed when the cli-server spawned them as children processes. On POSIX this was fine, as the I/O is asynchronous and the writes to the streams (not consumed, with content being buffered) didn't block. On Windows, however, where Node's pipes are blocking, once the buffers of the unconsumed streams filled up, blocked further execution.

We concluded that a reasonable fix was to perform the ipc over a separate file descriptor instead of stdin and explicitly ignoring or cosuming the rest for running managed scripts. That allows debug output to make it all the way down to the cli-server's output (and eventually do aggregated logging from a single point), but prevents tasks from getting blocked.