Fix handling of stderr and stdout preventing diffing to hang on Windows (#441)
* - api spec does not hang on launch anymore, so it can run in the foreground and subsequent commands can assume the daemon is started
- regression suite should verify that clis are handling output streams properly

* windows should use bash

* bash should only be used when necessary. It seems like using bash for the cargo steps does not work as intended

* node_modules should be cached per runner.os

* yarn should emit verbose information

* yarn should have a longer timeout in case CI is slow for large packages

* restore matrix

* yarn does not need to be verbose in the happy path

* scenarios should fail if they have not completed within 5 minutes

* focus on windows now

* diff engine stderr output should be consumed and written to the file system

* does the order matter?

* does the buffer size matter?

* github action should show useful output for debugging

* always show output

* explicit object mode, one at a time

* bring back other platforms to compare

* need more output to gauge worker progress

* limit concurrency

* highWaterMark experiment

* experiment: interactions buffer size

* Fix how runManagedScript leverages ipc, making sure stdout and stderr are consumed and non-blocking

* Remove highwatermarks from diff worker rust

* Revert changes to diff engine concurrency and buffer sizes

* Improve the DiffWorker for Rust's handling of logging output from the diff engine

* Make sure stderr and stdout are being consumed and don't block the diffing on Windows

The problem lied in that stdout and stderr streams not being consumed when the cli-server spawned them as children processes. On POSIX this was fine, as the I/O is asynchronous and the writes to the streams (not consumed, with content being buffered) didn't block. On Windows, however, where Node's pipes are blocking, once the buffers of the unconsumed streams filled up, blocked further execution.

We concluded that a reasonable fix was to perform the ipc over a separate file descriptor instead of stdin and explicitly ignoring or cosuming the rest for running managed scripts. That allows debug output to make it all the way down to the cli-server's output (and eventually do aggregated logging from a single point), but prevents tasks from getting blocked.

Co-authored-by: Dev Doshi <dev@useoptic.com>