file system capture saver should save all interactions. (#446)
* file system capture saver should save all interactions.
The interface ICaptureSaver has 3 methods: init(), save(interaction), and cleanup(). We are currently using it in a few different usage scenarios, so the interface doesn't provide any guarantees about whether a given interaction has been saved when save(interaction) resolves or when cleanup() resolves. It requires cooperation between the caller and the saver to achieve the desired behaviors. In addition, the batching logic is getting difficult to follow, test, and maintain. In the future we should consider redefining the interfaces and reimplementing (perhaps with xstate representing the parallel states). For now, when cleanup() is called on the file system capture saver, it should wait until all the incoming interactions have been saved and then stop listening for more batches of interactions. This approach has issues as well, since it doesn't stop accepting save(interaction) calls. Let's discuss

* restore e2e workflows

* use GITHUB_PATH instead of add-path

* use newer node setup action