replace shebang option with programmatic approach
Unfortunately, Linux (and Git Bash on windows) treats everything after the shebang interpreter as one option, while Mac treats each as separate. This makes it difficult to have more than one option. In lieu of that, we can disable the warning listener or use environment variables. Open to other solutions but for now we should make the build pass