Fix bug where split off spec changes pointing to a batch inside the root chunk couldn't be assembled.
Fixed by refactoring the HistoryProjection and making batch commits inside of the root chunk part of the commits graph as well. This can work by assuming that the parent of any batch commit _without_ a parent id is the batch commit that came before it in the event stream. During assembly, the first split-off spec change refers to the last batch commit found in the root chunk, instead of "root".