optic baseline: sniff-source Handle incomplete responses
In some cases, the HTTP response parser doesn't see an HTTP message as
complete.
This can happen if we don't process or miss bytes in the response.
Oddly, this also happens with simple `ab` invocations:
ab -c 1 -n 20 -H 'Accept-Encoding: gzip, deflate, br' http://127.0.0.1:3001/todos/

By this stage, we have most of the data ready and can actually process
it. We do this if the Simulation hasn't emitted anything and we have
something to emit.

Signed-off-by: Ray Bejjani <ray.bejjani@useoptic.com>