optic baseline: No sniff-source warnings on high-water-marks
These warnings caused some confusion, so they are gone. Node streams
continue to allocate memory and buffer data above this point.

Oddly, I can't cause this to happen on my machine. This could be a load
thing, where we don't drain the streams soon enough but I imagine
running chrome and ab on a laptop would be enough to hit this sometimes.

We can try to drops the data. Unfortuntely, where we expect to have
larger responses they are also more likely to be compressed. We can't
decompress the stream if we drop part of it so it would be easier to
drop the whole HTTP session outright then.

Signed-off-by: Ray Bejjani <ray.bejjani@useoptic.com>